+++ generators/app/templates/app/config/providers/urlRouterProvider.config.js
@@ -1,11 +1,5 @@
 /**
- * Generated header by <%= authorShort %> for <%= appNameKebab %> project
- * Generated file urlRouterProvider.config on WebStorm
- *
- * Created by: <%= authorShort %>
- * Date: <%= nowDate %>
- * Time: <%= nowTime %>
- * Version: 0.0.0
+ <%- include('../../../partial/header.txt') %>
  */
 (function (angular) {
 	'use strict';