+++ generators/app/templates/grunt/config/usemin.js
@@ -1,11 +1,5 @@
 /**
- * Generated header by C0ZEN for generator-cozen-angular project
- * Generated file usemin on WebStorm
- *
- * Created by: Geoffrey "C0ZEN" Testelin
- * Date: 19/11/2017
- * Time: 21:34
- * Version: 1.0.0
+ <%- include('../../partial/header.txt') %>
  */
 module.exports = {
 	html: [