+++ generators/app/templates/grunt/config/wiredep.js
@@ -1,11 +1,5 @@
 /**
- * Generated header by C0ZEN for generator-cozen-angular project
- * Generated file wiredep on WebStorm
- *
- * Created by: Geoffrey "C0ZEN" Testelin
- * Date: 19/11/2017
- * Time: 21:20
- * Version: 1.0.0
+ <%- include('../../partial/header.txt') %>
  */
 /* eslint key-spacing:"off" */
 module.exports = {